Added new Explore pages, with proper Header and Footer. Also added buttons and links to their proper pages.
