+++ src/App.js
@@ -10,9 +10,15 @@ import Profile from './pages/Profile/Profile';
 import FavoriteRecipes from './pages/FavoriteRecipes';
 import NotFoundPage from './pages/NotFoundPage';
 import RecipesMade from './pages/RecipesMade';
-import ExploreFoodsScreen from './components/ExploreFoods/ExploreFoodsScreen';
-import ExploreDrinksScreen from './components/ExploreDrinks/ExploreDrinksScreen';
+
 import Details from './components/Details/Details';
+
+import ExploreDrinksScreen from './pages/Explore/Drink/ExploreDrinks/ExploreDrinksScreen';
+
+import ExploreFoodsScreen from './pages/Explore/Food/ExploreFoods/ExploreFoodsScreen';
+import ExploreFoodByIng from './pages/Explore/Food/ExploreFoods/ExploreFoodByIng/ExploreFoodByIng';
+import ExploreFoodByArea from './pages/Explore/Food/ExploreFoods/ExploreFoodByArea/ExploreFoodByArea';
+import ExploreDrinkByIng from './pages/Explore/Drink/ExploreDrinks/ExploreDrinkByArea/ExploreDrinkByIng';
 
 function App() {
   return (
@@ -30,9 +36,9 @@ function App() {
             <Route exact path="/explorar" component={Explore} />
             <Route exact path="/explorar/comidas" component={ExploreFoodsScreen} />
             <Route exact path="/explorar/bebidas" component={ExploreDrinksScreen} />
-            <Route exact path="/explorar/comidas/ingredientes" component={null} />
-            <Route exact path="/explorar/bebidas/ingredientes" component={null} />
-            <Route exact path="/explorar/comidas/area" component={null} />
+            <Route exact path="/explorar/comidas/ingredientes" component={ExploreFoodByIng} />
+            <Route exact path="/explorar/bebidas/ingredientes" component={ExploreDrinkByIng} />
+            <Route exact path="/explorar/comidas/area" component={ExploreFoodByArea} />
             <Route exact path="/perfil" component={Profile} />
             <Route exact path="/receitas-feitas" component={RecipesMade} />
             <Route exact path="/receitas-favoritas" component={FavoriteRecipes} />
