--- src/components/ExploreDrinks/ExploreDrinksScreen.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import Header from '../Header/Header';
-import Footer from '../Footer/Footer';
-
-export default function ExploreDrinksScreen() {
-  return (
-    <div>
-      {Header('Explorar Bebidas', false)}
-      <Footer />
-    </div>
-  );
-}
