+++ src/components/FoodsAndDrinksDisplay/FoodsAndDrinksDisplay.style.css
@@ -1,11 +1,16 @@
 .container-display {
-  background-color: rgba(172, 148, 116, 0.418);
   border-bottom: 1px solid lightgray;
   border-radius: 15px;
+  box-shadow: 0 11px 18px -18px rgba(0, 0, 0, 0.795);
   display: flex;
   flex-direction: row;
   margin-bottom: 10px;
   padding: 15px;
+}
+
+.container-display:last-child {
+  border: none;
+  box-shadow: 0 11px 18px -18px rgba(0, 0, 0, 0.795);
 }
 
 .img-display {
@@ -14,7 +19,6 @@
 }
 
 .item-overflow {
-  box-shadow: 0 11px 18px -18px rgba(0, 0, 0, 0.795);
   display: flex;
   flex-direction: column;
   height: 410px;
