+++ src/pages/Explore/Food/ExploreFoods/ExploreFoodByArea/ExploreFoodByArea.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import Header from '../../../../../components/Header/Header';
+import Footer from '../../../../../components/Footer/Footer';
+
+export default function ExploreFoodByArea() {
+  return (
+    <div>
+      {Header('Explorar Origem', true)}
+      <Footer />
+    </div>
+  );
+}
