+++ src/pages/Explore/Food/ExploreFoods/ExploreFoodByIng/ExploreFoodByIng.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import Header from '../../../../../components/Header/Header';
+import Footer from '../../../../../components/Footer/Footer';
+
+export default function ExploreFoodByIng() {
+  return (
+    <div>
+      {Header('Explorar Ingredientes', false)}
+      <Footer />
+    </div>
+  );
+}
