+++ src/pages/Explore/Food/ExploreFoods/ExploreFoodsScreen.style.css
@@ -1,0 +1,15 @@
+.explore-btn-container {
+  align-items: center;
+  display: flex;
+}
+
+.explore-btn {
+  background: none;
+  border: none;
+  border-bottom: 1px solid lightgray;
+  color: rgb(70, 70, 70);
+  font-size: 15px;
+  height: 30px;
+  margin-bottom: 15px;
+  width: 150px;
+}
